refactor: Option 이름변경 및 패키지 이동
